Fix tests

+++ src/bitcoin_utxo_lp/types.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 from dataclasses import dataclass
-from decimal import Decimal
 from typing import Sequence
 
 
@@ -11,7 +10,7 @@ class UTXO:
     txid: str
     vout: int
     value_sats: int
-    input_vbytes: Decimal  # keep Decimal to avoid float surprises
+    input_vbytes: float
 
 
 @dataclass(frozen=True, slots=True)
@@ -24,9 +23,9 @@ class TxSizing:
     change_output_vbytes: vbytes for the change output
     """
 
-    base_overhead_vbytes: Decimal
-    recipient_output_vbytes: Decimal
-    change_output_vbytes: Decimal
+    base_overhead_vbytes: float
+    recipient_output_vbytes: float
+    change_output_vbytes: float
 
 
 @dataclass(frozen=True, slots=True)
@@ -34,7 +33,7 @@ class SelectionParams:
     """Fixed inputs for one coin-selection run."""
 
     target_sats: int
-    fee_rate_sat_per_vb: Decimal
+    fee_rate_sat_per_vb: float
     min_change_sats: int  # dust / wallet policy threshold
     sizing: TxSizing
 
@@ -58,5 +57,5 @@ class SelectionResult:
         return self.total_input_sats - self.fee_sats
 
 
-def total_input_vbytes(selected: Sequence[UTXO]) -> Decimal:
-    return sum((u.input_vbytes for u in selected), Decimal(0))
+def total_input_vbytes(selected: Sequence[UTXO]) -> float:
+    return sum((u.input_vbytes for u in selected), float(0))
